obico-ml-api-only/requirements.txt
click==8.1.7
requests==2.31.0
urllib3==2.1.0
Pillow==10.1.0
